refactor: Move test-utils out of lib/
