+++ lib/get-tsd-results.test.js
@@ -1,5 +1,5 @@
 import {getTsdResults} from './get-tsd-results'
-import {testHarnessFilePaths} from './test-utils'
+import {testHarnessFilePaths} from '../test-harness/test-utils'
 
 it('successfully compiles a passing type definition test', async () => {
   const result = await getTsdResults(testHarnessFilePaths.complicated)
