+++ test-harness/tests/describe-it-block-with-error.test-d.ts
