+++ test-harness/tests/describe-test-block-with-error.test-d.ts
