+++ test-harness/tests/has-failing-test.test-d.ts
