+++ test-harness/tests/it-block-with-name-error.test-d.ts
