+++ test-harness/tests/test-block-with-name-error.test-d.ts
